requirements.txt
streamlit
pillow
io
